% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_cost.R
\name{actual_cost}
\alias{actual_cost}
\title{Test computations of the cost}
\usage{
actual_cost(x, ca, pa, beta, betaP, type = c("meanvar", "mean"))
}
\arguments{
\item{x}{univariate time series}

\item{ca}{data frame of collective anomalies, must contain columns called start and end}

\item{pa}{data frame of point anomalies, must contain column called location}

\item{beta}{penalty for introducing a collective anomaly}

\item{betaP}{penalty for introducing a point anomaly}

\item{type}{type of gaussian anomaly}
}
\value{
total cost
}
\description{
Direct computation of the cost of a  partition for IID gaussian variables, with a background distribution of N(0,1)
}
\details{
The point anomaly evaluation does not include the correction term gamma
}
