% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{addBase,partition-method}
\alias{addBase,partition-method}
\title{Add a new baseline period}
\usage{
\S4method{addBase}{partition}(obj, f, b, t)
}
\arguments{
\item{obj}{partition object}

\item{f}{function to generate a new class segment}

\item{b}{penalty for introducing new segment}

\item{t}{time at start of segment}
}
\description{
Add a new baseline period
}
