% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{addPoint,partition-method}
\alias{addPoint,partition-method}
\title{Add a new Point anomaly}
\usage{
\S4method{addPoint}{partition}(obj, x, mu, sigma, f, b, t)
}
\arguments{
\item{obj}{partition object}

\item{mu}{mean of background distribution}

\item{sigma}{variance of background distribution}

\item{f}{function to generate a new class segment}

\item{b}{penalty for introducing new segment}

\item{t}{time at start of segment}
}
\description{
Add a new Point anomaly
}
