% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomalous-package.R
\docType{package}
\name{anomalous-package}
\alias{anomalous-package}
\title{Another package for anomoly and changepoint detection}
\description{
This offer a more more generic, but lower perfromance, implimentation of some of the algorithms in the `changepoint` and `anomoly` packages
}
