% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capa.R
\name{capa}
\alias{capa}
\title{An R implimentation of the segmented search algorithm}
\usage{
capa(part, prune = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{cost}{a cost function}
}
\value{
the optimal partition
}
\description{
An R implimentation of the segmented search algorithm
}
\details{
Basic R implimentation - not efficent
}
