% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capaP.R
\name{capaP}
\alias{capaP}
\title{An R implimentation of the pelt agorithm for partition structures}
\usage{
capaP(
  y,
  mu,
  sigma,
  anomType,
  beta,
  min_length = 2,
  max_length = .Machine$integer.max
)
}
\arguments{
\item{y}{univariate data series}

\item{mu}{background mean}

\item{sigma}{background variance}

\item{beta}{Penalisation term for each type of segment}

\item{min_length}{minimum length of a segment}

\item{max_length}{maximum length of a segment}

\item{segType}{defined types of segment allowed}
}
\description{
An R implimentation of the pelt agorithm for partition structures
}
\details{
Basic R implimentation of the PELT algorihm
}
