% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capaR.R
\name{capaR}
\alias{capaR}
\title{An R implimentation of the segmented search algorithm}
\usage{
capaR(
  y,
  mu,
  sigma,
  segType,
  beta,
  betaP,
  min_length = 2,
  max_length = .Machine$integer.max
)
}
\arguments{
\item{y}{univariate data series}

\item{mu}{mean of univariate series y}

\item{sigma}{variance of univariate series y}

\item{segType}{the names of the segment class to use}

\item{beta}{Penalisation term for a new segment}

\item{betaP}{Cost of new point anomaly}

\item{min_length}{minimum length of a segment}

\item{max_length}{maximum length of a segment}
}
\description{
An R implimentation of the segmented search algorithm
}
\details{
Basic R implimentation - not efficent
}
