% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_cost.R
\name{categoricalCost}
\alias{categoricalCost}
\title{R6 class for the categorical distribution}
\description{
Cost functions for the multinomial distribution
}
\details{
Collective anomalies are represented as chnages to the expected proportions.
Time varying expected proportions are currently not handled.
}
\examples{
set.seed(0)
m <- c(1:4)/sum(1:4)
X <- t(rmultinom(100, 1, m))

p <- categoricalCost$new(X,m)
p$baseCost(90,95) ## cost of non-anomalous distribution for x[90:95]
p$pointCost(90,0) ## point anomaly cost for x[90] with 0 penalty
## collective anomaly cost for x[90:95] with penalty of 57 and at least 3 observation
p$collectiveCost(90,95,57,3)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-categoricalCost-length}{\code{categoricalCost$length()}}
\item \href{#method-categoricalCost-new}{\code{categoricalCost$new()}}
\item \href{#method-categoricalCost-baseCost}{\code{categoricalCost$baseCost()}}
\item \href{#method-categoricalCost-pointCost}{\code{categoricalCost$pointCost()}}
\item \href{#method-categoricalCost-collectiveCost}{\code{categoricalCost$collectiveCost()}}
\item \href{#method-categoricalCost-param}{\code{categoricalCost$param()}}
\item \href{#method-categoricalCost-clone}{\code{categoricalCost$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-categoricalCost-length"></a>}}
\if{latex}{\out{\hypertarget{method-categoricalCost-length}{}}}
\subsection{Method \code{length()}}{
Get the length of time series
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{categoricalCost$length()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-categoricalCost-new"></a>}}
\if{latex}{\out{\hypertarget{method-categoricalCost-new}{}}}
\subsection{Method \code{new()}}{
Initialise the cost function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{categoricalCost$new(x, m = rep(1/ncol(x), ncol(x)))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{integer matrix of observations of 0,1}

\item{\code{m}}{numeric vector of expected proportions
## need to check row sums are 1 - possibly just use multinomial??}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-categoricalCost-baseCost"></a>}}
\if{latex}{\out{\hypertarget{method-categoricalCost-baseCost}{}}}
\subsection{Method \code{baseCost()}}{
Compute the non-anomalous cost of a segment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{categoricalCost$baseCost(a, b, pen = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{start of period}

\item{\code{b}}{end of period}

\item{\code{pen}}{penalty cost}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-categoricalCost-pointCost"></a>}}
\if{latex}{\out{\hypertarget{method-categoricalCost-pointCost}{}}}
\subsection{Method \code{pointCost()}}{
Compute the point anomaly cost of a time step
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{categoricalCost$pointCost(b, pen)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{b}}{time step}

\item{\code{pen}}{penalty cost}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-categoricalCost-collectiveCost"></a>}}
\if{latex}{\out{\hypertarget{method-categoricalCost-collectiveCost}{}}}
\subsection{Method \code{collectiveCost()}}{
Compute the anomalous cost of a segment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{categoricalCost$collectiveCost(a, b, pen, len)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{start of period}

\item{\code{b}}{end of period}

\item{\code{pen}}{penalty cost}

\item{\code{len}}{minimum number of observations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-categoricalCost-param"></a>}}
\if{latex}{\out{\hypertarget{method-categoricalCost-param}{}}}
\subsection{Method \code{param()}}{
Compute parameters of a segment if anomalous
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{categoricalCost$param(a, b)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{start of period}

\item{\code{b}}{end of period}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-categoricalCost-clone"></a>}}
\if{latex}{\out{\hypertarget{method-categoricalCost-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{categoricalCost$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
