% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{coef.anomalous_partition}
\alias{coef.anomalous_partition}
\title{Coefficent method for an amomalous_partition}
\usage{
\method{coef}{anomalous_partition}(object, ...)
}
\arguments{
\item{object}{the anomalous_partition}

\item{...}{optional parameters see details}
}
\value{
A matrix of parameters whose rows correspond to the rows in the summary output
}
\description{
Generates coefficents for the anomalous periods in an object with class amomalous_partition
}
\details{
Generates coefficents for all anomalous periods. Required input \code{cost} is a cost function. Optional input \code{t} is the time of the partitioning solution to use.
}
