% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{collective_anomalies}
\alias{collective_anomalies}
\title{Method for extracting collective anomalies}
\usage{
collective_anomalies(p, t)
}
\arguments{
\item{p}{object, such as a partition, from which to extract the point anomalies}

\item{t}{the end time at which the partition is based}
}
\description{
Extracts the collective anomaly summaries from the object
}
