% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{collective_change}
\alias{collective_change}
\title{return summary of Collective Changes}
\usage{
collective_change(p)
}
\arguments{
\item{p}{a partition object}
}
\description{
return summary of Collective Changes
}
