% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{cost,partition-method}
\alias{cost,partition-method}
\title{Cost of the partioning}
\usage{
\S4method{cost}{partition}(obj)
}
\arguments{
\item{obj}{partition object}
}
\description{
Cost of the partioning
}
