% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{createSegment}
\alias{createSegment}
\title{A wrapper function to create segments}
\usage{
createSegment(segType, beta, min_length, max_length)
}
\arguments{
\item{segType}{type of segment}

\item{beta}{Penalisation term for the segment}

\item{min_length}{minimum length of a segment}

\item{max_length}{maximum length of a segment}
}
\description{
A wrapper function to create segments
}
\details{
Basic R implimentation - not efficent
}
