% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crops.R
\name{crops}
\alias{crops}
\title{An implimentation of the CROPS algorithm in 1D}
\usage{
crops(
  betaMin,
  betaMax,
  fCost,
  alg = pelt,
  betaP = Inf,
  min_length = 2,
  prune = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{betaMin}{lower bound of penalisation window}

\item{betaMax}{upper bound of penalisation window}

\item{fCost}{the cost function}

\item{alg}{algorithm either capa of pelt}

\item{betaP}{penalty for adding a point anomaly - only for use with capa}

\item{min_length}{minimum number of values in a collective anomaly}

\item{prune}{logical, should pruning be used}

\item{verbose}{logical, print out progress}
}
\value{
something...
}
\description{
An implimentation of the CROPS algorithm in 1D
}
\details{
This will only work for cost functions where the beta is additive!!!
}
