% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss_cost.R
\name{gaussCost}
\alias{gaussCost}
\alias{gaussMean}
\alias{gaussVar}
\alias{gaussMeanVar}
\title{R6 class for Univariate Gaussian Cost Functions}
\arguments{
\item{x}{numeric vector or matrix of observations (see details)}

\item{m}{numeric vector of mean values}

\item{s}{numeric vector of standard deviation values}

\item{point_type}{representation of point anomalies as either a change in mean or variance}

\item{a}{start of period}

\item{b}{end of period}

\item{pen}{penalty cost}

\item{len}{minimum number of obseervations}
}
\description{
Cost functions for differening types of univariate gaussion anomalies
}
\details{
Collective anomalies are represented either as changes in mean (\code{gaussMean}),
variance (\code{gaussVar}) or mean and variance (\code{gaussMeanvar}). See vignettes for details.
If x is a matrix then the values in each row are treated as IID replicate observations.
}
\examples{
set.seed(0)
m <- runif(100)
s <- pmax(1e-4,runif(100))
x <- rnorm(100,m,s) ## example data

gM <- gaussMean$new(x,m,s) ## anomalies are changes in mean
gM$baseCost(90,95) ## cost of non-anomalous distribution for x[90:95]
gM$pointCost(90,0) ## point anomaly cost for x[90] with 0 penalty
## collective anomaly cost for x[90:95] with penalty of 57 and at least 3 observation
gM$collectiveCost(90,95,57,3)

gV <- gaussVar$new(x,m,s) ## anomalies are changes in variance
gV$baseCost(90,95) ## cost of non-anomalous distribution for x[90:95]
gV$pointCost(90,0) ## point anomaly cost for x[90] with 0 penalty
## collective anomaly cost for x[90:95] with penalty of 57 and at least 3 observation
gV$collectiveCost(90,95,57,3)

gMV <- gaussMeanVar$new(x,m,s) ## anomalies are changes in mean and variance
gMV$baseCost(90,95) ## cost of non-anomalous distribution for x[90:95]
gMV$pointCost(90,0) ## point anomaly cost for x[90] with 0 penalty
## collective anomaly cost for x[90:95] with penalty of 57 and at least 3 observation
gMV$collectiveCost(90,95,57,3)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-gaussCost-length}{\code{gaussCost$length()}}
\item \href{#method-gaussCost-new}{\code{gaussCost$new()}}
\item \href{#method-gaussCost-baseCost}{\code{gaussCost$baseCost()}}
\item \href{#method-gaussCost-pointCost}{\code{gaussCost$pointCost()}}
\item \href{#method-gaussCost-clone}{\code{gaussCost$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gaussCost-length"></a>}}
\if{latex}{\out{\hypertarget{method-gaussCost-length}{}}}
\subsection{Method \code{length()}}{
Get the length of time series
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gaussCost$length()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gaussCost-new"></a>}}
\if{latex}{\out{\hypertarget{method-gaussCost-new}{}}}
\subsection{Method \code{new()}}{
Initialise the cost function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gaussCost$new(x, m = 0, s = 1, point_type = c("var", "mean"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{numeric vector of observations}

\item{\code{m}}{numeric vector of mean values}

\item{\code{s}}{numeric vector of standard deviation values}

\item{\code{point_type}}{representation of point anomalies as either a change in mean or variance}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gaussCost-baseCost"></a>}}
\if{latex}{\out{\hypertarget{method-gaussCost-baseCost}{}}}
\subsection{Method \code{baseCost()}}{
Compute the non-anomalous cost of a segment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gaussCost$baseCost(a, b, pen = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{start of period}

\item{\code{b}}{end of period}

\item{\code{pen}}{penalty cost}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gaussCost-pointCost"></a>}}
\if{latex}{\out{\hypertarget{method-gaussCost-pointCost}{}}}
\subsection{Method \code{pointCost()}}{
Compute the point anomaly cost of a time step
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gaussCost$pointCost(b, pen)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{b}}{time step}

\item{\code{pen}}{penalty cost}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gaussCost-clone"></a>}}
\if{latex}{\out{\hypertarget{method-gaussCost-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gaussCost$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{\link[anomalous:gaussCost]{anomalous::gaussCost}} -> \code{gaussMean}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-gaussMean-collectiveCost}{\code{gaussMean$collectiveCost()}}
\item \href{#method-gaussMean-param}{\code{gaussMean$param()}}
\item \href{#method-gaussMean-clone}{\code{gaussMean$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="anomalous" data-topic="gaussCost" data-id="baseCost"><a href='../../anomalous/html/gaussCost.html#method-gaussCost-baseCost'><code>anomalous::gaussCost$baseCost()</code></a></span></li>
<li><span class="pkg-link" data-pkg="anomalous" data-topic="gaussCost" data-id="initialize"><a href='../../anomalous/html/gaussCost.html#method-gaussCost-initialize'><code>anomalous::gaussCost$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="anomalous" data-topic="gaussCost" data-id="length"><a href='../../anomalous/html/gaussCost.html#method-gaussCost-length'><code>anomalous::gaussCost$length()</code></a></span></li>
<li><span class="pkg-link" data-pkg="anomalous" data-topic="gaussCost" data-id="pointCost"><a href='../../anomalous/html/gaussCost.html#method-gaussCost-pointCost'><code>anomalous::gaussCost$pointCost()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gaussMean-collectiveCost"></a>}}
\if{latex}{\out{\hypertarget{method-gaussMean-collectiveCost}{}}}
\subsection{Method \code{collectiveCost()}}{
Compute the anomalous cost of a segment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gaussMean$collectiveCost(a, b, pen, len)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{start of period}

\item{\code{b}}{end of period}

\item{\code{pen}}{penalty cost}

\item{\code{len}}{minimum number of observations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gaussMean-param"></a>}}
\if{latex}{\out{\hypertarget{method-gaussMean-param}{}}}
\subsection{Method \code{param()}}{
Compute parameters of a segment if anomalous
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gaussMean$param(a, b)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{start of period}

\item{\code{b}}{end of period}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gaussMean-clone"></a>}}
\if{latex}{\out{\hypertarget{method-gaussMean-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gaussMean$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{\link[anomalous:gaussCost]{anomalous::gaussCost}} -> \code{gaussVar}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-gaussVar-collectiveCost}{\code{gaussVar$collectiveCost()}}
\item \href{#method-gaussVar-param}{\code{gaussVar$param()}}
\item \href{#method-gaussVar-clone}{\code{gaussVar$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="anomalous" data-topic="gaussCost" data-id="baseCost"><a href='../../anomalous/html/gaussCost.html#method-gaussCost-baseCost'><code>anomalous::gaussCost$baseCost()</code></a></span></li>
<li><span class="pkg-link" data-pkg="anomalous" data-topic="gaussCost" data-id="initialize"><a href='../../anomalous/html/gaussCost.html#method-gaussCost-initialize'><code>anomalous::gaussCost$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="anomalous" data-topic="gaussCost" data-id="length"><a href='../../anomalous/html/gaussCost.html#method-gaussCost-length'><code>anomalous::gaussCost$length()</code></a></span></li>
<li><span class="pkg-link" data-pkg="anomalous" data-topic="gaussCost" data-id="pointCost"><a href='../../anomalous/html/gaussCost.html#method-gaussCost-pointCost'><code>anomalous::gaussCost$pointCost()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gaussVar-collectiveCost"></a>}}
\if{latex}{\out{\hypertarget{method-gaussVar-collectiveCost}{}}}
\subsection{Method \code{collectiveCost()}}{
Compute the anomalous cost of a segment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gaussVar$collectiveCost(a, b, pen, len)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{start of period}

\item{\code{b}}{end of period}

\item{\code{pen}}{penalty cost}

\item{\code{len}}{minimum number of observations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gaussVar-param"></a>}}
\if{latex}{\out{\hypertarget{method-gaussVar-param}{}}}
\subsection{Method \code{param()}}{
Compute parameters of a segment if anomalous
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gaussVar$param(a, b)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{start of period}

\item{\code{b}}{end of period}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gaussVar-clone"></a>}}
\if{latex}{\out{\hypertarget{method-gaussVar-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gaussVar$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{\link[anomalous:gaussCost]{anomalous::gaussCost}} -> \code{gaussMeanVar}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-gaussMeanVar-collectiveCost}{\code{gaussMeanVar$collectiveCost()}}
\item \href{#method-gaussMeanVar-param}{\code{gaussMeanVar$param()}}
\item \href{#method-gaussMeanVar-clone}{\code{gaussMeanVar$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="anomalous" data-topic="gaussCost" data-id="baseCost"><a href='../../anomalous/html/gaussCost.html#method-gaussCost-baseCost'><code>anomalous::gaussCost$baseCost()</code></a></span></li>
<li><span class="pkg-link" data-pkg="anomalous" data-topic="gaussCost" data-id="initialize"><a href='../../anomalous/html/gaussCost.html#method-gaussCost-initialize'><code>anomalous::gaussCost$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="anomalous" data-topic="gaussCost" data-id="length"><a href='../../anomalous/html/gaussCost.html#method-gaussCost-length'><code>anomalous::gaussCost$length()</code></a></span></li>
<li><span class="pkg-link" data-pkg="anomalous" data-topic="gaussCost" data-id="pointCost"><a href='../../anomalous/html/gaussCost.html#method-gaussCost-pointCost'><code>anomalous::gaussCost$pointCost()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gaussMeanVar-collectiveCost"></a>}}
\if{latex}{\out{\hypertarget{method-gaussMeanVar-collectiveCost}{}}}
\subsection{Method \code{collectiveCost()}}{
Compute the non-anomalous cost of a segment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gaussMeanVar$collectiveCost(a, b, pen, len)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{start of period}

\item{\code{b}}{end of period}

\item{\code{pen}}{penalty cost}

\item{\code{len}}{minimum number of observations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gaussMeanVar-param"></a>}}
\if{latex}{\out{\hypertarget{method-gaussMeanVar-param}{}}}
\subsection{Method \code{param()}}{
Compute parameters of a segment if anomalous
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gaussMeanVar$param(a, b)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{start of period}

\item{\code{b}}{end of period}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gaussMeanVar-clone"></a>}}
\if{latex}{\out{\hypertarget{method-gaussMeanVar-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gaussMeanVar$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
