% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussSegment.R
\name{gaussKnown}
\alias{gaussKnown}
\title{create a gaussKnown segment}
\usage{
gaussKnown(beta, start)
}
\arguments{
\item{beta}{penalty term}

\item{start}{time step at start}
}
\description{
create a gaussKnown segment
}
