% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss_cost.R
\docType{class}
\name{gaussMean-class}
\alias{gaussMean-class}
\title{Gaussian Cost Function for a univariate series}
\description{
Gaussian Cost Function for a univariate series
}
\section{Slots}{

\describe{
\item{\code{summaryStats}}{a record of the summary statistics required, one row per time step}

\item{\code{maxT}}{length of the series}
}}

