% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussSegment.R
\name{gaussMean}
\alias{gaussMean}
\title{create a gaussMean segment}
\usage{
gaussMean(beta, start)
}
\arguments{
\item{beta}{penalty term}

\item{start}{time step at start}
}
\description{
create a gaussMean segment
}
