% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussSegment.R
\name{gaussMeanVar}
\alias{gaussMeanVar}
\title{gaussMeanVar}
\usage{
gaussMeanVar(beta, t)

gaussMeanVar(beta, t)
}
\arguments{
\item{pen}{Penalisation term for the segment}
}
\description{
gaussMeanVar

gaussMeanVar
}
