% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{gaussPartition}
\alias{gaussPartition}
\title{gaussPartition}
\usage{
gaussPartition(beta, min_length, max_length)
}
\arguments{
\item{beta}{Named vector of penalisation terms for starting a new segment}

\item{min_length}{minimum length of a segment}

\item{max_length}{maximum length of a segment}
}
\description{
gaussPartition
}
\details{
Basic R implimentation - not efficent
}
