% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussSegment.R
\name{gaussPoint}
\alias{gaussPoint}
\title{create a gaussPoint segment}
\usage{
gaussPoint(beta, start)
}
\arguments{
\item{beta}{penalty term}

\item{start}{time step at start}
}
\description{
create a gaussPoint segment
}
