% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss_reg.R
\name{gaussRegCost}
\alias{gaussRegCost}
\alias{gaussRegMean}
\alias{gaussRegVar}
\alias{gaussRegMeanVar}
\title{R6 class for Univariate Gaussian Cost Functions}
\description{
Cost functions for differening types of univariate gaussion anomalies
}
\details{
Collective anomalies are represented either as changes in parameters describing the mean (\code{gaussRegMean}),
variance (\code{gaussRegVar}) or mean and variance (\code{gaussRegMeanVar}). Changes in variance are represented as a scaling parameter, not changes to the covariance. See vignettes for details.

Each element of the input \code{x} should be a list containing a vector of observations \code{y} and corresponding design matrix \code{X}. Optionally in can also include a vector of parameter \code{m} and covariance matrix \code{S}.
}
\examples{
## simple test
set.seed(10)
x <- list()
n <- 120
for(ii in 1:48){
    
    if(ii < 10){ theta = c(1,0); sigma <- 0.1 }
    if(ii >= 10 & ii <12){ theta <- c(10,0); sigma <- 2}
    if(ii >= 12 & ii < 44){ theta <- c(5,1); sigma <- 2}
    if(ii >= 44 ){ theta <- c(1,0); sigma <- 0.1}
    
    X <- cbind(rep(1,n),runif(n,ii-1,ii))
    y <- rnorm(n, X\%*\%theta, sigma)
    x[[ii]] <- list(y=y,X=X)
}

fCost <- gaussRegMeanVar$new(x)
p <- partition(4*log(sum(sapply(y,length))),NA,2)
res <- pelt(p,fCost)

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-gaussRegCost-length}{\code{gaussRegCost$length()}}
\item \href{#method-gaussRegCost-new}{\code{gaussRegCost$new()}}
\item \href{#method-gaussRegCost-baseCost}{\code{gaussRegCost$baseCost()}}
\item \href{#method-gaussRegCost-pointCost}{\code{gaussRegCost$pointCost()}}
\item \href{#method-gaussRegCost-clone}{\code{gaussRegCost$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gaussRegCost-length"></a>}}
\if{latex}{\out{\hypertarget{method-gaussRegCost-length}{}}}
\subsection{Method \code{length()}}{
Get the length of time series
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gaussRegCost$length()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gaussRegCost-new"></a>}}
\if{latex}{\out{\hypertarget{method-gaussRegCost-new}{}}}
\subsection{Method \code{new()}}{
Initialise the cost function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gaussRegCost$new(x, non_neg = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{a list of regressions (see details)}

\item{\code{non_neg}}{should only non-negative paraemter solutions be considered}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gaussRegCost-baseCost"></a>}}
\if{latex}{\out{\hypertarget{method-gaussRegCost-baseCost}{}}}
\subsection{Method \code{baseCost()}}{
Compute the non-anomalous cost of a segment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gaussRegCost$baseCost(a, b, pen = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{start of period}

\item{\code{b}}{end of period}

\item{\code{pen}}{penalty cost}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gaussRegCost-pointCost"></a>}}
\if{latex}{\out{\hypertarget{method-gaussRegCost-pointCost}{}}}
\subsection{Method \code{pointCost()}}{
Compute the point anomaly cost of a time step
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gaussRegCost$pointCost(b, pen)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{b}}{time step}

\item{\code{pen}}{penalty cost}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gaussRegCost-clone"></a>}}
\if{latex}{\out{\hypertarget{method-gaussRegCost-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gaussRegCost$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{\link[anomalous:gaussRegCost]{anomalous::gaussRegCost}} -> \code{gaussRegMean}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-gaussRegMean-collectiveCost}{\code{gaussRegMean$collectiveCost()}}
\item \href{#method-gaussRegMean-param}{\code{gaussRegMean$param()}}
\item \href{#method-gaussRegMean-clone}{\code{gaussRegMean$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="anomalous" data-topic="gaussRegCost" data-id="baseCost"><a href='../../anomalous/html/gaussRegCost.html#method-gaussRegCost-baseCost'><code>anomalous::gaussRegCost$baseCost()</code></a></span></li>
<li><span class="pkg-link" data-pkg="anomalous" data-topic="gaussRegCost" data-id="initialize"><a href='../../anomalous/html/gaussRegCost.html#method-gaussRegCost-initialize'><code>anomalous::gaussRegCost$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="anomalous" data-topic="gaussRegCost" data-id="length"><a href='../../anomalous/html/gaussRegCost.html#method-gaussRegCost-length'><code>anomalous::gaussRegCost$length()</code></a></span></li>
<li><span class="pkg-link" data-pkg="anomalous" data-topic="gaussRegCost" data-id="pointCost"><a href='../../anomalous/html/gaussRegCost.html#method-gaussRegCost-pointCost'><code>anomalous::gaussRegCost$pointCost()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gaussRegMean-collectiveCost"></a>}}
\if{latex}{\out{\hypertarget{method-gaussRegMean-collectiveCost}{}}}
\subsection{Method \code{collectiveCost()}}{
Compute the anomalous cost of a segment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gaussRegMean$collectiveCost(a, b, pen, len)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{start of period}

\item{\code{b}}{end of period}

\item{\code{pen}}{penalty cost}

\item{\code{len}}{minimum number of observations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gaussRegMean-param"></a>}}
\if{latex}{\out{\hypertarget{method-gaussRegMean-param}{}}}
\subsection{Method \code{param()}}{
Compute parameters of a segment if anomalous
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gaussRegMean$param(a, b)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{start of period}

\item{\code{b}}{end of period}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gaussRegMean-clone"></a>}}
\if{latex}{\out{\hypertarget{method-gaussRegMean-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gaussRegMean$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{\link[anomalous:gaussRegCost]{anomalous::gaussRegCost}} -> \code{gaussRegVar}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-gaussRegVar-collectiveCost}{\code{gaussRegVar$collectiveCost()}}
\item \href{#method-gaussRegVar-param}{\code{gaussRegVar$param()}}
\item \href{#method-gaussRegVar-clone}{\code{gaussRegVar$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="anomalous" data-topic="gaussRegCost" data-id="baseCost"><a href='../../anomalous/html/gaussRegCost.html#method-gaussRegCost-baseCost'><code>anomalous::gaussRegCost$baseCost()</code></a></span></li>
<li><span class="pkg-link" data-pkg="anomalous" data-topic="gaussRegCost" data-id="initialize"><a href='../../anomalous/html/gaussRegCost.html#method-gaussRegCost-initialize'><code>anomalous::gaussRegCost$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="anomalous" data-topic="gaussRegCost" data-id="length"><a href='../../anomalous/html/gaussRegCost.html#method-gaussRegCost-length'><code>anomalous::gaussRegCost$length()</code></a></span></li>
<li><span class="pkg-link" data-pkg="anomalous" data-topic="gaussRegCost" data-id="pointCost"><a href='../../anomalous/html/gaussRegCost.html#method-gaussRegCost-pointCost'><code>anomalous::gaussRegCost$pointCost()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gaussRegVar-collectiveCost"></a>}}
\if{latex}{\out{\hypertarget{method-gaussRegVar-collectiveCost}{}}}
\subsection{Method \code{collectiveCost()}}{
Compute the anomalous cost of a segment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gaussRegVar$collectiveCost(a, b, pen, len)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{start of period}

\item{\code{b}}{end of period}

\item{\code{pen}}{penalty cost}

\item{\code{len}}{minimum number of observations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gaussRegVar-param"></a>}}
\if{latex}{\out{\hypertarget{method-gaussRegVar-param}{}}}
\subsection{Method \code{param()}}{
Compute parameters of a segment if anomalous
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gaussRegVar$param(a, b)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{start of period}

\item{\code{b}}{end of period}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gaussRegVar-clone"></a>}}
\if{latex}{\out{\hypertarget{method-gaussRegVar-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gaussRegVar$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{\link[anomalous:gaussRegCost]{anomalous::gaussRegCost}} -> \code{gaussRegMeanVar}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-gaussRegMeanVar-collectiveCost}{\code{gaussRegMeanVar$collectiveCost()}}
\item \href{#method-gaussRegMeanVar-param}{\code{gaussRegMeanVar$param()}}
\item \href{#method-gaussRegMeanVar-clone}{\code{gaussRegMeanVar$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="anomalous" data-topic="gaussRegCost" data-id="baseCost"><a href='../../anomalous/html/gaussRegCost.html#method-gaussRegCost-baseCost'><code>anomalous::gaussRegCost$baseCost()</code></a></span></li>
<li><span class="pkg-link" data-pkg="anomalous" data-topic="gaussRegCost" data-id="initialize"><a href='../../anomalous/html/gaussRegCost.html#method-gaussRegCost-initialize'><code>anomalous::gaussRegCost$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="anomalous" data-topic="gaussRegCost" data-id="length"><a href='../../anomalous/html/gaussRegCost.html#method-gaussRegCost-length'><code>anomalous::gaussRegCost$length()</code></a></span></li>
<li><span class="pkg-link" data-pkg="anomalous" data-topic="gaussRegCost" data-id="pointCost"><a href='../../anomalous/html/gaussRegCost.html#method-gaussRegCost-pointCost'><code>anomalous::gaussRegCost$pointCost()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gaussRegMeanVar-collectiveCost"></a>}}
\if{latex}{\out{\hypertarget{method-gaussRegMeanVar-collectiveCost}{}}}
\subsection{Method \code{collectiveCost()}}{
Compute the anomalous cost of a segment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gaussRegMeanVar$collectiveCost(a, b, pen, len)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{start of period}

\item{\code{b}}{end of period}

\item{\code{pen}}{penalty cost}

\item{\code{len}}{minimum number of observations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gaussRegMeanVar-param"></a>}}
\if{latex}{\out{\hypertarget{method-gaussRegMeanVar-param}{}}}
\subsection{Method \code{param()}}{
Compute parameters of a segment if anomalous
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gaussRegMeanVar$param(a, b)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{start of period}

\item{\code{b}}{end of period}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gaussRegMeanVar-clone"></a>}}
\if{latex}{\out{\hypertarget{method-gaussRegMeanVar-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gaussRegMeanVar$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
