% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guass_mean.R
\name{GuassianMean}
\alias{GuassianMean}
\alias{guass_mean}
\title{Univariate Gaussian Mean estimation functions}
\usage{
guass_mean(obj, y, grp, pen)
}
\arguments{
\item{y}{new data point}

\item{pen}{penality value}

\item{s}{data frame of summary statistics}
}
\description{
Basic R implimentation of functions for estimating the guassian mean
}
\details{
`guass_mean_update` Add new observation to summary statistics
}
