% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lad_cost.R
\name{ladCost}
\alias{ladCost}
\title{R6 class for LAD Cost Functions}
\description{
Cost functions for the Least Absolution Deviation from a given quantile
}
\details{
this is a very niaive and slow implimentation
}
\examples{
set.seed(0)
m <- runif(100)
x <- rnorm(100,m)

p <- ladCost$new(x,m,0.5)
p$baseCost(90,95) ## cost of non-anomalous distribution for x[90:95]
p$pointCost(90,0) ## point anomaly cost for x[90] with 0 penalty
## collective anomaly cost for x[90:95] with penalty of 57 and at least 3 observation
p$collectiveCost(90,95,57,3)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ladCost-length}{\code{ladCost$length()}}
\item \href{#method-ladCost-new}{\code{ladCost$new()}}
\item \href{#method-ladCost-baseCost}{\code{ladCost$baseCost()}}
\item \href{#method-ladCost-pointCost}{\code{ladCost$pointCost()}}
\item \href{#method-ladCost-collectiveCost}{\code{ladCost$collectiveCost()}}
\item \href{#method-ladCost-param}{\code{ladCost$param()}}
\item \href{#method-ladCost-clone}{\code{ladCost$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ladCost-length"></a>}}
\if{latex}{\out{\hypertarget{method-ladCost-length}{}}}
\subsection{Method \code{length()}}{
Get the length of time series
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ladCost$length()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ladCost-new"></a>}}
\if{latex}{\out{\hypertarget{method-ladCost-new}{}}}
\subsection{Method \code{new()}}{
Initialise the cost function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ladCost$new(x, m = 0, tau = 0.5)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{numeric vector of observations}

\item{\code{m}}{expected value of x}

\item{\code{tau}}{the quantile}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ladCost-baseCost"></a>}}
\if{latex}{\out{\hypertarget{method-ladCost-baseCost}{}}}
\subsection{Method \code{baseCost()}}{
Compute the non-anomalous cost of a segment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ladCost$baseCost(a, b, pen = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{start of period}

\item{\code{b}}{end of period}

\item{\code{pen}}{penalty cost}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ladCost-pointCost"></a>}}
\if{latex}{\out{\hypertarget{method-ladCost-pointCost}{}}}
\subsection{Method \code{pointCost()}}{
Compute the point anomaly cost of a time step
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ladCost$pointCost(a, pen)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{time step}

\item{\code{pen}}{penalty cost}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ladCost-collectiveCost"></a>}}
\if{latex}{\out{\hypertarget{method-ladCost-collectiveCost}{}}}
\subsection{Method \code{collectiveCost()}}{
Compute the anomalous cost of a segment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ladCost$collectiveCost(a, b, pen, len)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{start of period}

\item{\code{b}}{end of period}

\item{\code{pen}}{penalty cost}

\item{\code{len}}{minimum number of observations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ladCost-param"></a>}}
\if{latex}{\out{\hypertarget{method-ladCost-param}{}}}
\subsection{Method \code{param()}}{
Compute parameters of a segment if anomalous
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ladCost$param(a, b)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{start of period}

\item{\code{b}}{end of period}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ladCost-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ladCost-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ladCost$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
