% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_cost.R
\name{multinomialCost}
\alias{multinomialCost}
\title{R6 class for the multinomial distribution}
\description{
Cost functions for the multinomial distribution
}
\details{
Collective anomalies are represented as chnages to the expected proportions.
Time varying expected proportions are currently not handled.
}
\examples{
set.seed(0)
m <- c(1:4)/sum(1:4)
X <- t(rmultinom(100, 144, m))

p <- multinomialCost$new(X,m)
p$baseCost(90,95) ## cost of non-anomalous distribution for x[90:95]
p$pointCost(90,0) ## point anomaly cost for x[90] with 0 penalty
## collective anomaly cost for x[90:95] with penalty of 57 and at least 3 observation
p$collectiveCost(90,95,57,3)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-multinomialCost-length}{\code{multinomialCost$length()}}
\item \href{#method-multinomialCost-new}{\code{multinomialCost$new()}}
\item \href{#method-multinomialCost-baseCost}{\code{multinomialCost$baseCost()}}
\item \href{#method-multinomialCost-pointCost}{\code{multinomialCost$pointCost()}}
\item \href{#method-multinomialCost-collectiveCost}{\code{multinomialCost$collectiveCost()}}
\item \href{#method-multinomialCost-param}{\code{multinomialCost$param()}}
\item \href{#method-multinomialCost-clone}{\code{multinomialCost$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-multinomialCost-length"></a>}}
\if{latex}{\out{\hypertarget{method-multinomialCost-length}{}}}
\subsection{Method \code{length()}}{
Get the length of time series
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{multinomialCost$length()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-multinomialCost-new"></a>}}
\if{latex}{\out{\hypertarget{method-multinomialCost-new}{}}}
\subsection{Method \code{new()}}{
Initialise the cost function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{multinomialCost$new(x, m = rep(1/ncol(x), ncol(x)))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{integer matrix of observations}

\item{\code{m}}{numeric vector of expected proportions}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-multinomialCost-baseCost"></a>}}
\if{latex}{\out{\hypertarget{method-multinomialCost-baseCost}{}}}
\subsection{Method \code{baseCost()}}{
Compute the non-anomalous cost of a segment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{multinomialCost$baseCost(a, b, pen = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{start of period}

\item{\code{b}}{end of period}

\item{\code{pen}}{penalty cost}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-multinomialCost-pointCost"></a>}}
\if{latex}{\out{\hypertarget{method-multinomialCost-pointCost}{}}}
\subsection{Method \code{pointCost()}}{
Compute the point anomaly cost of a time step
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{multinomialCost$pointCost(b, pen)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{b}}{time step}

\item{\code{pen}}{penalty cost}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-multinomialCost-collectiveCost"></a>}}
\if{latex}{\out{\hypertarget{method-multinomialCost-collectiveCost}{}}}
\subsection{Method \code{collectiveCost()}}{
Compute the anomalous cost of a segment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{multinomialCost$collectiveCost(a, b, pen, len)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{start of period}

\item{\code{b}}{end of period}

\item{\code{pen}}{penalty cost}

\item{\code{len}}{minimum number of observations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-multinomialCost-param"></a>}}
\if{latex}{\out{\hypertarget{method-multinomialCost-param}{}}}
\subsection{Method \code{param()}}{
Compute parameters of a segment if anomalous
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{multinomialCost$param(a, b)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{start of period}

\item{\code{b}}{end of period}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-multinomialCost-clone"></a>}}
\if{latex}{\out{\hypertarget{method-multinomialCost-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{multinomialCost$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
