% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op.R
\name{op}
\alias{op}
\title{An R implimentation of the segmented search algorithm}
\usage{
op(y, fC, Beta)
}
\arguments{
\item{y}{univariate data series}

\item{fC}{cost function}

\item{Beta}{Penalisation term}

\item{Kappa}{Pruning term}
}
\description{
An R implimentation of the segmented search algorithm
}
\details{
Basic R implimentation - not efficent
}
