% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{partition}
\alias{partition}
\title{Partion record for the data}
\usage{
partition(beta, betaP, min_length)
}
\arguments{
\item{beta}{penalty for a new collective anomaly}

\item{betaP}{penalty for a new point anomaly}

\item{min_length}{shortest length of a collective anomaly}
}
\description{
A partition records the seperation of the data generated by the pelt or capa methods
}
\examples{
p <- partition(3,4,2)
}
