% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelt.R
\name{pelt}
\alias{pelt}
\title{An R implimentation of the pelt agorithm}
\usage{
pelt(
  y,
  mu,
  sigma,
  segType,
  beta,
  min_length = 2,
  max_length = .Machine$integer.max
)
}
\arguments{
\item{y}{univariate data series}

\item{segType}{function to generate an object of class Segment}

\item{beta}{Penalisation term}

\item{min_length}{minimum length of a segment}

\item{max_length}{maximum length of a segment}
}
\description{
An R implimentation of the pelt agorithm
}
\details{
Basic R implimentation of the PELT algorihm
}
