% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{plot.anomalous_partition}
\alias{plot.anomalous_partition}
\title{Plot method for an amomalous_partition}
\usage{
\method{plot}{anomalous_partition}(x, ...)
}
\arguments{
\item{x}{the anomalous_partition}

\item{...}{optional parameters see details}
}
\value{
Produces a plot
}
\description{
Provides a summary plot of an object with class amomalous_partition
}
\details{
If providing \code{t} a time the results prodiced are as though the analysis was stopped then.
The optional inputs \code{xx} and \code{yy} allow for data to be displayed under the shaded anomaly/changepoint areas.
If no data is provided \code{xx = 1:t} and \code{yy} is the cost of the periods the timestep is in.
}
