% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{point_anomalies}
\alias{point_anomalies}
\title{return summary of point anomalies}
\usage{
point_anomalies(p)
}
\arguments{
\item{p}{a partition object}
}
\description{
return summary of point anomalies
}
