% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_cost.R
\name{poisCost}
\alias{poisCost}
\title{R6 class for Univariate Poisson Cost Functions}
\description{
Cost functions for the univariate Poission distribution
}
\details{
Collective anomalies are represented as multiplicative changes in rate
}
\examples{
set.seed(0)
r <- 8 + runif(100)*2
x <- rpois(100,lambda = r)

p <- poisCost$new(x,r)
p$baseCost(90,95) ## cost of non-anomalous distribution for x[90:95]
p$pointCost(90,0) ## point anomaly cost for x[90] with 0 penalty
## collective anomaly cost for x[90:95] with penalty of 57 and at least 3 observation
p$collectiveCost(90,95,57,3) 
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-poisCost-length}{\code{poisCost$length()}}
\item \href{#method-poisCost-new}{\code{poisCost$new()}}
\item \href{#method-poisCost-baseCost}{\code{poisCost$baseCost()}}
\item \href{#method-poisCost-pointCost}{\code{poisCost$pointCost()}}
\item \href{#method-poisCost-collectiveCost}{\code{poisCost$collectiveCost()}}
\item \href{#method-poisCost-param}{\code{poisCost$param()}}
\item \href{#method-poisCost-clone}{\code{poisCost$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-poisCost-length"></a>}}
\if{latex}{\out{\hypertarget{method-poisCost-length}{}}}
\subsection{Method \code{length()}}{
Get the length of time series
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{poisCost$length()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-poisCost-new"></a>}}
\if{latex}{\out{\hypertarget{method-poisCost-new}{}}}
\subsection{Method \code{new()}}{
Initialise the cost function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{poisCost$new(x, rate = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{numeric vector of observations}

\item{\code{rate}}{numeric vector of rate parameters}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-poisCost-baseCost"></a>}}
\if{latex}{\out{\hypertarget{method-poisCost-baseCost}{}}}
\subsection{Method \code{baseCost()}}{
Compute the non-anomalous cost of a segment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{poisCost$baseCost(a, b, pen = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{start of period}

\item{\code{b}}{end of period}

\item{\code{pen}}{penalty cost}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-poisCost-pointCost"></a>}}
\if{latex}{\out{\hypertarget{method-poisCost-pointCost}{}}}
\subsection{Method \code{pointCost()}}{
Compute the point anomaly cost of a time step
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{poisCost$pointCost(b, pen)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{b}}{time step}

\item{\code{pen}}{penalty cost}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-poisCost-collectiveCost"></a>}}
\if{latex}{\out{\hypertarget{method-poisCost-collectiveCost}{}}}
\subsection{Method \code{collectiveCost()}}{
Compute the anomalous cost of a segment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{poisCost$collectiveCost(a, b, pen, len)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{start of period}

\item{\code{b}}{end of period}

\item{\code{pen}}{penalty cost}

\item{\code{len}}{minimum number of observations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-poisCost-param"></a>}}
\if{latex}{\out{\hypertarget{method-poisCost-param}{}}}
\subsection{Method \code{param()}}{
Compute parameters of a segment if anomalous
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{poisCost$param(a, b)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{start of period}

\item{\code{b}}{end of period}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-poisCost-clone"></a>}}
\if{latex}{\out{\hypertarget{method-poisCost-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{poisCost$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
