% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_cost.R
\name{rankCost}
\alias{rankCost}
\title{R6 class for rank based anomalies}
\description{
This is VERY developmental - do not use
}
\examples{
set.seed(0)
m <- runif(100)
s <- pmax(1e-4,runif(100))
x <- rnorm(100,m,s) ## example data
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-rankCost-length}{\code{rankCost$length()}}
\item \href{#method-rankCost-new}{\code{rankCost$new()}}
\item \href{#method-rankCost-baseCost}{\code{rankCost$baseCost()}}
\item \href{#method-rankCost-pointCost}{\code{rankCost$pointCost()}}
\item \href{#method-rankCost-collectiveCost}{\code{rankCost$collectiveCost()}}
\item \href{#method-rankCost-param}{\code{rankCost$param()}}
\item \href{#method-rankCost-clone}{\code{rankCost$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rankCost-length"></a>}}
\if{latex}{\out{\hypertarget{method-rankCost-length}{}}}
\subsection{Method \code{length()}}{
Get the length of time series
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rankCost$length()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rankCost-new"></a>}}
\if{latex}{\out{\hypertarget{method-rankCost-new}{}}}
\subsection{Method \code{new()}}{
Initialise the cost function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rankCost$new(x, m = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{numeric matrix of observations}

\item{\code{m}}{numeric vector or matrix of mean values}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rankCost-baseCost"></a>}}
\if{latex}{\out{\hypertarget{method-rankCost-baseCost}{}}}
\subsection{Method \code{baseCost()}}{
Compute the non-anomalous cost of a segment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rankCost$baseCost(a, b, pen = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{start of period}

\item{\code{b}}{end of period}

\item{\code{pen}}{penalty cost}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rankCost-pointCost"></a>}}
\if{latex}{\out{\hypertarget{method-rankCost-pointCost}{}}}
\subsection{Method \code{pointCost()}}{
Compute the point anomaly cost of a time step
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rankCost$pointCost(b, pen)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{b}}{time step}

\item{\code{pen}}{penalty cost}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rankCost-collectiveCost"></a>}}
\if{latex}{\out{\hypertarget{method-rankCost-collectiveCost}{}}}
\subsection{Method \code{collectiveCost()}}{
Compute the non-anomalous cost of a segment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rankCost$collectiveCost(a, b, pen, len)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{start of period}

\item{\code{b}}{end of period}

\item{\code{pen}}{penalty cost}

\item{\code{len}}{minimum number of observations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rankCost-param"></a>}}
\if{latex}{\out{\hypertarget{method-rankCost-param}{}}}
\subsection{Method \code{param()}}{
Compute parameters of a segment if anomalous
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rankCost$param(a, b)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{start of period}

\item{\code{b}}{end of period}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rankCost-clone"></a>}}
\if{latex}{\out{\hypertarget{method-rankCost-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rankCost$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
