% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{summary.anomalous_partition}
\alias{summary.anomalous_partition}
\title{Summmary method for an amomalous_partition}
\usage{
\method{summary}{anomalous_partition}(object, ...)
}
\arguments{
\item{object}{the anomalous_partition}

\item{...}{optional parameters see details}
}
\value{
A data.frame summarising the partitions
}
\description{
Provides a summary of an object with class amomalous_partition
}
\details{
If providing \code{t} a time the results prodiced are as though the analysis was stopped then.
}
