% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costs.R, R/partition.R
\name{UnivariateCost}
\alias{UnivariateCost}
\alias{fCmn}
\alias{fCv}
\alias{fCmnv}
\alias{fCpois}
\title{Univariate cost functions}
\usage{
fCmn(y, cl, beta)

fCv(y, cl, beta)

fCmnv(y, cl, beta)

fCpois(y, cl, beta)

fCmn(y, cl, beta)

fCv(y, cl, beta)

fCmnv(y, cl, beta)

fCpois(y, cl, beta)
}
\arguments{
\item{y}{univariate data series}
}
\description{
Basic R implimentation of various univariate cost function

Basic R implimentation of various univariate cost function
}
\details{
`fCmn` Gaussian mean, common known variance

`fCv` Gaussian variance, common known mean

`fCmnv` Gaussian mean and variance

`fCpois` Poisson

`fCmn` Gaussian mean, common known variance

`fCv` Gaussian variance, common known mean

`fCmnv` Gaussian mean and variance

`fCpois` Poisson
}
