% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{update}
\alias{update}
\alias{update,meanSegment-method}
\alias{update,varSegment-method}
\alias{update,meanvarSegment-method}
\alias{update,poisSegment-method}
\title{Update a segment}
\usage{
update(obj, x)

\S4method{update}{meanSegment}(obj, x)

\S4method{update}{varSegment}(obj, x)

\S4method{update}{meanvarSegment}(obj, x)

\S4method{update}{poisSegment}(obj, x)
}
\arguments{
\item{obj}{a segment class object}

\item{x}{new observations to add}
}
\description{
Update a segment
}
