% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{update,partition-method}
\alias{update,partition-method}
\title{Update a partion object}
\usage{
\S4method{update}{partition}(obj, x, mu, sigma)
}
\arguments{
\item{obj}{partition object}

\item{x}{observed value}

\item{mu}{mean of background distribution}

\item{sigma}{variance of background distribution}

\item{isPoint}{add data to the most recent point}
}
\description{
Update a partion object
}
