% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{zerooneSegment}
\alias{zerooneSegment}
\title{zerooneSegment}
\usage{
zerooneSegment(x, beta, min_length, max_length)
}
\arguments{
\item{x}{initial data}

\item{beta}{Penalisation term for the segment}

\item{min_length}{minimum length of a segment}

\item{max_length}{maximum length of a segment}
}
\description{
zerooneSegment
}
\details{
Basic R implimentation - not efficent
}
