% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomalous-package.R
\docType{package}
\name{anomalous-package}
\alias{anomalous}
\alias{anomalous-package}
\alias{dynatop-package}
\title{anomalous: What the package does (short line)}
\description{
More about what it does (maybe more than one line).
}
\keyword{internal}
