% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capa.R
\name{capa}
\alias{capa}
\title{An R implimentation of the segmented search algorithm}
\usage{
capa(
  y,
  mu,
  sigma,
  segType,
  beta,
  betaP,
  min_length = 2,
  max_length = .Machine$integer.max
)
}
\arguments{
\item{y}{univariate data series}

\item{segType}{function to generate an object of class Segment}

\item{beta}{Penalisation term for a new segment}

\item{min_length}{minimum length of a segment}

\item{max_length}{maximum length of a segment}
}
\description{
An R implimentation of the segmented search algorithm
}
\details{
Basic R implimentation - not efficent
}
