% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capa.R
\name{capa}
\alias{capa}
\title{An R implimentation of the segmented search algorithm}
\usage{
capa(part, fCost, prune = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{part}{the starting partition}

\item{fCost}{the cost function}

\item{prune}{logical, should pruning be used}

\item{verbose}{logical, print out progress}
}
\value{
the optimal partition
}
\description{
An R implimentation of the segmented search algorithm
}
\details{
Basic R implimentation of capa - not efficent
}
