% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{collective_anomalies}
\alias{collective_anomalies}
\title{return summary of Collective Changes}
\usage{
collective_anomalies(p, type = NULL)
}
\arguments{
\item{p}{a partition object}
}
\description{
return summary of Collective Changes
}
