% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussSegment.R
\name{gaussMeanVar}
\alias{gaussMeanVar}
\title{create a gaussMeanVar segment}
\usage{
gaussMeanVar(beta, start)
}
\arguments{
\item{beta}{penalty term}

\item{start}{time step at start}
}
\description{
create a gaussMeanVar segment
}
