% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussSegment.R
\name{gaussPoint}
\alias{gaussPoint}
\title{gaussPoint}
\usage{
gaussPoint(beta, t)
}
\arguments{
\item{beta}{Penalisation term for the segment}

\item{t}{start time of segment}
}
\description{
gaussPoint
}
