% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\name{param}
\alias{param}
\title{Get the parameters for a partitioning result}
\usage{
param(res, fCost)
}
\arguments{
\item{res}{the result of a partitioning algorithm}

\item{fCost}{the cost function}
}
\value{
list of parameters
}
\description{
Get the parameters for a partitioning result
}
\details{
Not yet implimented for all cost functions
}
