% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelt.R
\name{pelt}
\alias{pelt}
\title{An R implimentation of the pelt agorithm}
\usage{
pelt(y, fC, beta)
}
\arguments{
\item{y}{univariate data series}

\item{fC}{cost function}

\item{Beta}{Penalisation term}

\item{min_length}{minimum length of a segment}
}
\description{
An R implimentation of the pelt agorithm
}
\details{
Basic r implimentation of the PELT algorihm
}
