% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelt.R
\name{pelt}
\alias{pelt}
\title{An R implimentation of the segmented search algorithmpelt algorithm}
\usage{
pelt(part, fCost, prune = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{part}{the starting partition}

\item{fCost}{the cost function}

\item{prune}{logical, should pruning be used}

\item{verbose}{logical, print out progress}
}
\value{
the optimal partition
}
\description{
An R implimentation of the segmented search algorithmpelt algorithm
}
\details{
Basic R implimentation of pelt - not efficent
}
