% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{point_anomalies}
\alias{point_anomalies}
\title{Method for extracting point anomalies}
\usage{
point_anomalies(p, t)
}
\arguments{
\item{p}{object, such as a partition, from which to extract the point anomalies}

\item{t}{the end time at which the partition is based}
}
\description{
Extracts the point anomaly summaries from the object
}
