% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costs.R
\name{UnivariateCost}
\alias{UnivariateCost}
\alias{fCmn}
\alias{fCv}
\alias{fCmnv}
\title{Univariate cost functions}
\usage{
fCmn(y)

fCv(y)

fCmnv(y)
}
\arguments{
\item{y}{univariate data series}
}
\description{
Basic R implimentation of various univariate cost function
}
\details{
`fCmn` Gaussian mean, common known variance

`fCv` Gaussian variance, common known mean

`fCmnv` Gaussian mean and variance
}
