% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/segment.R
\name{update}
\alias{update}
\alias{update,gaussMean-method}
\alias{update,gaussVar-method}
\alias{update,gaussMeanVar-method}
\alias{update,poisRate-method}
\title{Update a segment with new observations}
\usage{
update(obj, x, ...)

\S4method{update}{gaussMean}(obj, x, mu, sigma)

\S4method{update}{gaussVar}(obj, x, mu, sigma)

\S4method{update}{gaussMeanVar}(obj, x, mu, sigma)

\S4method{update}{poisRate}(obj, x, mu, sigma)
}
\arguments{
\item{obj}{a segment class object}

\item{x}{new observations to add}
}
\description{
Update a segment with new observations
}
