% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{update}
\alias{update}
\alias{addPoint}
\alias{addCollective}
\title{Update a segment with new observations}
\usage{
update(obj, x, ...)

addPoint(obj, ...)

addCollective(obj, ...)
}
\arguments{
\item{obj}{a segment class object}

\item{x}{new observations to add}

\item{...}{other inputs}
}
\description{
Update a segment with new observations

Add a point segment

Add a point segment
}
